% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{auroc}
\alias{auroc}
\title{AUC for sNPLS-DA model}
\usage{
auroc(object)
}
\arguments{
\item{object}{A sNPLS object}
}
\value{
The area under the ROC curve for the model
}
\description{
AUC for a sNPLS-DA model
}
