% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread-data.R
\docType{data}
\name{bread}
\alias{bread}
\title{Bread data}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(bread)
}
\description{
Evaluation of ten bread with respect to eleven attributes
by eight judges (Xbread). The outcome is the salt content
of each bread (Ybread).
}
\references{
Bro, R, Multi-way Analysis in the Food Industry.
Models, Algorithms, and Applications. 1998.  PhD thesis,
University of Amsterdam (NL) & Royal Veterinary and Agricultural
University (DK).
}
\keyword{datasets}
