% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{coef.sNPLS}
\alias{coef.sNPLS}
\title{Coefficients from a sNPLS model}
\usage{
\method{coef}{sNPLS}(object, as.matrix = FALSE, ...)
}
\arguments{
\item{object}{A sNPLS model fit}

\item{as.matrix}{Should the coefficients be presented as matrix or vector?}

\item{...}{Further arguments passed to \code{coef}}
}
\value{
A matrix (or vector) of coefficients
}
\description{
Extract coefficients from a sNPLS model
}

