% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{cv_snpls}
\alias{cv_snpls}
\title{Cross-validation for a sNPLS model}
\usage{
cv_snpls(X_npls, Y_npls, ncomp = 1:3, keepJ = 1:ncol(X_npls),
  keepK = 1:dim(X_npls)[3], nfold = 10, parallel = TRUE, free_cores = 2)
}
\arguments{
\item{X_npls}{A three-way array containing the predictors.}

\item{Y_npls}{A matrix containing the response.}

\item{ncomp}{A vector with the different number of components to test}

\item{keepJ}{A vector with the different number of selected variables to test}

\item{keepK}{A vector with the different number of selected 'times' to test}

\item{nfold}{Number of folds for the cross-validation}

\item{parallel}{Should the computations be performed in parallel?}

\item{free_cores}{If parallel computations are performed how many cores are left unused}
}
\value{
A list with the best parameters for the model and the CV error
}
\description{
Performs cross-validation for a sNPLS model
}
\examples{
\dontrun{
X_npls<-array(rpois(7500, 10), dim=c(50, 50, 3))

Y_npls<-matrix(2+0.4*X_npls[,5,1]+0.7*X_npls[,10,1]-0.9*X_npls[,15,1]+
0.6*X_npls[,20,1]- 0.5*X_npls[,25,1]+rnorm(50), ncol=1)

cv1<- cv_snpls(X_npls, Y_npls, ncomp=1:2, keepJ = 1:3, keepK = 1:2, parallel = FALSE)
}
}
