% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{fitted.sNPLS}
\alias{fitted.sNPLS}
\title{Fitted method for sNPLS models}
\usage{
\method{fitted}{sNPLS}(object, ...)
}
\arguments{
\item{object}{A sNPLS model fit}

\item{...}{Further arguments passed to \code{fitted}}
}
\value{
Fitted values for the sNPLS model
}
\description{
Fitted method for sNPLS models
}
