% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{ga_snpls}
\alias{ga_snpls}
\title{Genetic Algorithm for selection of hyperparameter values}
\usage{
ga_snpls(
  X,
  Y,
  ncomp = c(1, 3),
  threshold_j = c(0, 1),
  threshold_k = c(0, 1),
  maxiter = 20,
  popSize = 50,
  parallel = TRUE,
  replicates = 10,
  metric = "RMSE",
  method = "sNPLS",
  ...
)
}
\arguments{
\item{X}{A three-way array containing the predictors.}

\item{Y}{A matrix containing the response.}

\item{ncomp}{A vector with the minimum and maximum number of components to assess}

\item{threshold_j}{Vector with threshold min and max values on Wj. Scaled between [0, 1)}

\item{threshold_k}{Vector with threshold min and max values on Wk. Scaled between [0, 1)}

\item{maxiter}{Maximum number of iterations (generations) of the genetic algorithm}

\item{popSize}{Population size (see \code{GA::ga()} documentation)}

\item{parallel}{Should the computations be performed in parallel? (see \code{GA::ga()} documentation)}

\item{replicates}{Number of replicates for the cross-validation performed in the fitness function of the genetic algoritm}

\item{metric}{Select between RMSE or AUC (for 0/1 response)}

\item{method}{Select between sNPLS, sNPLS-SR or sNPLS-VIP}

\item{...}{Further arguments passed to \code{GA::ga()}}
}
\value{
A summary of the genetic algorithm results
}
\description{
Runs a genetic algorithm to select the best combination of hyperparameter values
}
