% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot.cvsNPLS}
\alias{plot.cvsNPLS}
\title{Plot cross validation results for sNPLS objects}
\usage{
\method{plot}{cvsNPLS}(x, ...)
}
\arguments{
\item{x}{A cv_sNPLS object}

\item{...}{Arguments passed to \code{car::scatter3d}}
}
\value{
A 3D scatter plot with the results of the cross validation
}
\description{
Plot function for visualization of cross validation results for sNPLS models
}
