% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot.repeatcv}
\alias{plot.repeatcv}
\title{Density plot for repeat_cv results}
\usage{
\method{plot}{repeatcv}(x, ...)
}
\arguments{
\item{x}{A repeatcv object}

\item{...}{Further arguments passed to plot}
}
\value{
A grid of slices from a 3-D density plot of the results of the repeated cross-validation
}
\description{
Plots a grid of slices from the 3-D kernel denity estimates of the repeat_cv function
}
