% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot.sNPLS}
\alias{plot.sNPLS}
\title{Plots for sNPLS model fits}
\usage{
\method{plot}{sNPLS}(x, type = "T", comps = c(1, 2), labels = TRUE, group = NULL, ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{type}{The type of plot. One of those: "T", "U", "Wj", "Wk", "time" or "variables"}

\item{comps}{Vector with the components to plot. It can be of length \code{ncomp} for types "time" and "variables" and of length 2 otherwise.}

\item{labels}{Should rownames be added as labels to the plot?}

\item{group}{Vector with categorical variable defining groups (optional)}

\item{...}{Not used}
}
\value{
A plot of the type specified in the \code{type} parameter
}
\description{
Different plots for sNPLS model fits
}
