% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot_variables}
\alias{plot_variables}
\title{Internal function for \code{plot.sNPLS}}
\usage{
plot_variables(x, comps, xlab = "Variables", ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{comps}{A vector with the components to plot}

\item{xlab}{X-axis label}

\item{...}{Options passed to \code{plot}}
}
\value{
A plot of Wj coefficients for each component
}
\description{
Internal function for \code{plot.sNPLS}
}
