% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot_Wk}
\alias{plot_Wk}
\title{Internal function for \code{plot.sNPLS}}
\usage{
plot_Wk(x, comps, ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{comps}{A vector of length two with the components to plot}

\item{...}{Options passed to \code{plot}}
}
\value{
A plot of the Wk coefficients
}
\description{
Internal function for \code{plot.sNPLS}
}
