% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{Rmatrix}
\alias{Rmatrix}
\title{R-matrix from a sNPLS model fit}
\usage{
Rmatrix(x)
}
\arguments{
\item{x}{A sNPLS model obtained from \code{sNPLS}}
}
\value{
Returns the R-matrix of the model, needed to compute the coefficients
}
\description{
Builds the R-matrix from a sNPLS model fit
}
