% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{SR}
\alias{SR}
\title{Compute Selectivity Ratio for a sNPLS model}
\usage{
SR(model)
}
\arguments{
\item{model}{A sNPLS model}
}
\value{
A list of data.frames, each of them including the computed Selectivity Ratios for each variable
}
\description{
Estimates Selectivity Ratio for the different components of a sNPLS model fit
}
