% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{summary.sNPLS}
\alias{summary.sNPLS}
\title{Summary for sNPLS models}
\usage{
\method{summary}{sNPLS}(object, ...)
}
\arguments{
\item{object}{A sNPLS object}

\item{...}{Further arguments passed to summary.default}
}
\value{
A summary inclunding number of components, squared error and coefficients of the fitted model
}
\description{
Summary of a sNPLS model fit
}
