% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{unfold3w}
\alias{unfold3w}
\title{Unfolding of three-way arrays}
\usage{
unfold3w(x)
}
\arguments{
\item{x}{A three-way array}
}
\value{
Returns a matrix with dimensions \code{dim(x)[1] x dim(x)[2]*dim(x([3]))}
}
\description{
Unfolds a three-way array into a matrix
}

