% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{cv_fit}
\alias{cv_fit}
\title{Internal function for \code{cv_snpls}}
\usage{
cv_fit(xtrain, ytrain, xval, yval, ncomp, keepJ, keepK)
}
\arguments{
\item{xtrain}{A three-way training array}

\item{ytrain}{A response training matrix}

\item{xval}{A three-way test array}

\item{yval}{A response test matrix}

\item{ncomp}{Number of components for the sNPLS model}

\item{keepJ}{Number of variables to keep for each component}

\item{keepK}{Number of 'times' to keep for each component}
}
\value{
Returns the CV mean squared error
}
\description{
Internal function for \code{cv_snpls}
}
