% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot.sNPLS}
\alias{plot.sNPLS}
\title{Plots for sNPLS model fits}
\usage{
\method{plot}{sNPLS}(x, comps = c(1, 2), type = "T", ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{comps}{A vector of length two with the components to plot}

\item{type}{The type of plot. One of those: "T", "U", "Wj", "Wk" or "time"}

\item{...}{Options passed to \code{plot}}
}
\value{
A plot of the type specified in the \code{type} parameter
}
\description{
Different plots for sNPLS model fits
}

