% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot_time}
\alias{plot_time}
\title{Internal function for \code{plot.sNPLS}}
\usage{
plot_time(x, ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{...}{Options passed to \code{plot}}
}
\value{
A plot of Wk coefficients for each component
}
\description{
Internal function for \code{plot.sNPLS}
}
