% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNPLS_fit.R
\name{plot_U}
\alias{plot_U}
\title{Internal function for \code{plot.sNPLS}}
\usage{
plot_U(x, comps, ylim = c(min(x$U[, comps[2]]) - diff(range(x$U[,
  comps[2]]))/10, max(x$U[, comps[2]]) + diff(range(x$U[, comps[2]]))/10),
  xlim = c(min(x$U[, comps[1]]) - diff(range(x$U[, comps[1]]))/10, max(x$U[,
  comps[1]]) + diff(range(x$U[, comps[1]]))/10), ...)
}
\arguments{
\item{x}{A sNPLS model fit}

\item{comps}{A vector of length two with the components to plot}

\item{ylim}{Limits of the Y axis}

\item{xlim}{Limits of the X axis}

\item{...}{Options passed to \code{plot}}
}
\value{
A plot of the U matrix of a sNPLS model fit
}
\description{
Internal function for \code{plot.sNPLS}
}
